% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PackageDataAsset}
\alias{PackageDataAsset}
\title{PackageDataAsset}
\format{
An object of class \code{data.frame} with 614 rows and 11 columns.
}
\usage{
PackageDataAsset
}
\description{
PackageDataAsset
}
\keyword{datasets}
